🚚 move src to cams and bin to scripts
