+++ cams/cli.py
