+++ pyproject.toml
@@ -7,8 +7,8 @@ extend-select = ["F401"]  # F401 = unused-import
 
 ## Run options
 include = [
-  "src",
-  "bin",
+  "cams",
+  "scripts",
   "tests",
 ]
 ignore = []
@@ -86,7 +86,7 @@ reportUnknownMemberType = "error"
 reportMissingParameterType = "error"
 reportMissingTypeArgument = "error"
 
-reportCallInDefaultInitializer = "warning"
+reportCallInDefaultInitializer = "none"
 
 reportUnnecessaryIsInstance = "error"
 reportUnnecessaryCast = "error"
