+++ scripts/main.py
@@ -1,26 +1,20 @@
 """Script used to interact directly with the lightning cli.
 It is recommended to use the specialized scripts to interact with your models,
 for better checkpoint management.
-- Use `bin/fit_and_val.py` to fit and validate from scratch or a checkpoint.
-- Use `bin/predict.py` to run inference from a checkpoint.
+- Use `scripts/fit_and_val.py` to fit and validate from scratch or a checkpoint.
+- Use `scripts/predict.py` to run inference from a checkpoint.
 """
 
 from lightning.pytorch.core import LightningDataModule
 
-from src.cli import CamsCli
-from src.datamodule import CamsDataModule
-from src.plmodule import CamsLightningModule
+from cams.datamodule import CamsDataModule
 
 
 def cli_main(
     datamodule: type[LightningDataModule] = CamsDataModule,
     args: list[str] = None,
 ) -> None:
-    cli = CamsCli(
-        model_class=CamsLightningModule,
-        datamodule_class=datamodule,
-        args=args,
-    )
+    raise NotImplementedError()
 
 
 if __name__ == "__main__":
